Refactoring ModeComponent to be more autonomous
